pyproject.toml
[tool.poetry]
name = "features"
version = "0.1.0"
description = "Babylon Features & RAG Pipeline"
authors = ["Alan Ponte <alanjponte@gmail.com>"]
readme = "README.md"
#packages = [{ include = "your_package_name" }]  # or remove if not using a package dir

[tool.poetry.dependencies]
python = ">=3.13.5, <3.14"

requests = "^2.32.5"
pymongo = "4.15.1"
langchain-huggingface = "^0.3.1"
langchain-chroma = "^0.2.6"
langchain-core = "^0.3.79"
langchain-community = "^0.3.31"
# Secrets manager
hvac = "2.3.0"
sentence-transformers = "^5.1.1"
langchain-qdrant = "0.2.0"
python-dotenv = "^1.2.1"
langchain-text-splitters = "^0.3.0"
zenml = "^0.93.3"
datasets = "^4.5.0"
google-genai = "^1.65.0"

[tool.poetry.group.test.dependencies]
black = "26.1.0"
mypy = "^1.17.1"
pylint = "3.3.7"
pytest="^8.4.1"

# Dependencies for data visualization.
[tool.poetry.group.visualize.dependencies]
numpy = "2.3.4"
matplotlib = "^3.10.7"
plotly = "^6.3.1"
scikit-learn = "1.7.2"

[tool.poetry.group.harness.dependencies]
# Evaluating loguru as logging handler.
# See https://github.com/Delgan/loguru/releases
loguru="^0.7.3"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
